refactor: rename image files for course and semester pages to follow naming conventions
